src/main/java/ua/foxminded/javaspring/kocherga/web_application/service/UserService.java
package ua.foxminded.javaspring.kocherga.web_application.service;

import jakarta.transaction.Transactional;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import ua.foxminded.javaspring.kocherga.web_application.models.User;
import ua.foxminded.javaspring.kocherga.web_application.repository.UserRepository;

import java.util.List;

@Service
public class UserService {

    private final UserRepository userRepository;

    @Autowired
    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Transactional
    public List<User> getUsersByGroupId(int groupId) {
        return userRepository.findByOwnerGroupId(groupId);
    }

    @Transactional
    public List<User> getAllUsers() {
        return userRepository.findAll();
    }
}
